Backend/requirements.txt
Django
psycopg2-binary
Pillow
django-versatileimagefield
django-environ
djangorestframework
python-magic-bin
django-utils-six
djangorestframework-simplejwt
graphene_django
django-countries